refactor(git): isolate refresh orchestration
Extract the git diff refresh state machine and the shell refresh-forwarding seam into focused hooks so the behavior stays easier to reason about and test without changing refresh semantics.
